Refactor functions
