+++ functions/src/callables.ts
@@ -1,0 +1,9 @@
+import * as functions from 'firebase-functions';
+import { auth } from "./utils/admin";
+
+export const getUsersByEmail = functions.https.onCall((data, context) => {
+  if (!data.emails) {
+    return [];
+  }
+  return Promise.all(data.emails.map((email: string) => auth.getUserByEmail(email)));
+});
