+++ functions/src/utils/sendEmail.ts
